Continued organizing bug branch working dir
