--- errorPage.jsx
@@ -1,10 +1,0 @@
-import React, { Component } from 'react';
-
-class Error extends Component {
-    state = {  }
-    render() { 
-        return ( <h1>An error occured</h1> );
-    }
-}
- 
-export default Error;
