+++ my-app/src/scenarioPalpation.jsx
@@ -1,0 +1,10 @@
+import React, { Component } from 'react';
+
+class NewPalpation extends Component {
+    state = {  }
+    render() { 
+        return ( <h3>TPalpation</h3> );
+    }
+}
+ 
+export default NewPalpation;
