--- scenarioSpecial.jsx
@@ -1,10 +1,0 @@
-import React, { Component } from 'react';
-
-class NewSpecial extends Component {
-    state = {  }
-    render() { 
-        return ( <h5>Special</h5> );
-    }
-}
- 
-export default NewSpecial;
